+++ straxen/plugins/led_calibration.py
@@ -18,10 +18,10 @@ channel_list = [i for i in range(494)]
                  default=(0,40),
                  help="Window (samples) for baseline calculation."),
     strax.Option('led_window',
-                 default=(75, 105),
+                 default=(78, 116),
                  help="Window (samples) where we expect the signal in LED calibration"),
     strax.Option('noise_window',
-                 default=(20, 50),
+                 default=(10, 48),
                  help="Window (samples) to analysis the noise"),
     strax.Option('channel_list',
                  default=(tuple(channel_list)),
@@ -29,12 +29,17 @@ channel_list = [i for i in range(494)]
 
 class LEDCalibration(strax.Plugin):
     """
-    Preliminary version, several parameters to set during commisioning.
-    LEDCalibration returns: channel, time, dt, lenght, Area, amplitudeLED and amplitudeNOISE.
+    Preliminary version, several parameters to set during commissioning.
+    LEDCalibration returns: channel, time, dt, lenght, Area,
+    amplitudeLED and amplitudeNOISE.
     The new variables are:
-    - Area: Area computed in the given window, averaged over 6 windows that have the same starting sample and different end samples.
-    - amplitudeLED: peak amplitude of the LED on run in the given window.
-    - amplitudeNOISE: amplitude of the LED on run in a window far from the signal one.
+        - Area: Area computed in the given window, averaged over 6
+        windowsthat have the same starting sample and different end
+        samples.
+        - amplitudeLED: peak amplitude of the LED on run in the given
+        window.
+        - amplitudeNOISE: amplitude of the LED on run in a window far
+         from the signal one.
     """
     
     __version__ = '0.2.3'
@@ -61,42 +66,39 @@ class LEDCalibration(strax.Plugin):
         rr   = raw_records[mask]
         r    = get_records(rr, baseline_window=self.config['baseline_window'])
         del rr, raw_records
-               
+
         temp = np.zeros(len(r), dtype=self.dtype)
-        
+
         temp['channel'] = r['channel']
         temp['time']    = r['time']
         temp['dt']      = r['dt']
         temp['length']  = r['length']
-        
+
         on, off = get_amplitude(r, self.config['led_window'], self.config['noise_window'])
         temp['amplitude_led']   = on['amplitude']
         temp['amplitude_noise'] = off['amplitude']
-               
+
         area = get_area(r, self.config['led_window'])
         temp['area'] = area['area']
-        
         return temp
-    
+
+
 def get_records(raw_records, baseline_window):
     """
     Determine baseline as the average of the first baseline_samples
     of each pulse. Subtract the pulse float(data) from baseline.
-    """  
-    
-    if len(raw_records):
-        record_length = len(raw_records['data'][0])
-    else:
-        return
-        
-    _dtype = [(('Start time since unix epoch [ns]', 'time'), '<i8'), 
-              (('Length of the interval in samples', 'length'), '<i4'), 
-              (('Width of one sample [ns]', 'dt'), '<i2'), 
-              (('Channel/PMT number', 'channel'), '<i2'), 
-              (('Length of pulse to which the record belongs (without zero-padding)', 'pulse_length'), '<i4'), 
-              (('Fragment number in the pulse', 'record_i'), '<i2'), 
+    """
+
+    record_length = np.shape(raw_records.dtype['data'])[0]
+
+    _dtype = [(('Start time since unix epoch [ns]', 'time'), '<i8'),
+              (('Length of the interval in samples', 'length'), '<i4'),
+              (('Width of one sample [ns]', 'dt'), '<i2'),
+              (('Channel/PMT number', 'channel'), '<i2'),
+              (('Length of pulse to which the record belongs (without zero-padding)', 'pulse_length'), '<i4'),
+              (('Fragment number in the pulse', 'record_i'), '<i2'),
               (('Waveform data in raw ADC counts', 'data'), 'f4', (record_length,))]
-              
+
     records = np.zeros(len(raw_records), dtype=_dtype)
 
     records['time']         = raw_records['time']