+++ straxen/plugins/peak_processing.py
@@ -9,13 +9,18 @@ import strax
 import straxen
 from straxen.common import pax_file, get_resource, first_sr1_run
 export, __all__ = strax.exporter()
-
+from .pulse_processing import  HE_PREAMBLE
 
 @export
 @strax.takes_config(
     strax.Option('n_top_pmts', default=straxen.n_top_pmts,
                  help="Number of top PMTs"))
 class PeakBasics(strax.Plugin):
+    """
+    Compute the basic peak-properties, thereby dropping structured
+    arrays.
+    NB: This plugin can therefore be loaded as a pandas DataFrame.
+    """
     __version__ = "0.0.7"
     parallel = True
     depends_on = ('peaks',)
@@ -75,7 +80,7 @@ class PeakBasics(strax.Plugin):
         m = p['area'] > 0
         r['area_fraction_top'][m] = area_top[m]/p['area'][m]
         r['area_fraction_top'][~m] = float('nan')
-        r['rise_time'] = -p['area_decile_from_midpoint'][:,1]
+        r['rise_time'] = -p['area_decile_from_midpoint'][:, 1]
         return r
 
     @staticmethod
@@ -88,6 +93,18 @@ class PeakBasics(strax.Plugin):
                 t += t_i * p['dt'] * weight
             result[p_i] = t / p['area']
         return result
+
+
+@export
+class PeakBasicsHighEnergy(PeakBasics):
+    __doc__ = HE_PREAMBLE + PeakBasics.__doc__
+    __version__ = '0.0.1'
+    depends_on = 'peaks_he'
+    provides = 'peak_basics_he'
+    child_ends_with = '_he'
+
+    def compute(self, peaks_he):
+        return super().compute(peaks_he)
 
 
 @export
@@ -111,6 +128,7 @@ class PeakBasics(strax.Plugin):
                  help="Number of top PMTs")
 )
 class PeakPositions(strax.Plugin):
+    """Compute the S2 (x,y)-position based on a neural net."""
     dtype = [('x', np.float32,
               'Reconstructed S2 X position (cm), uncorrected'),
              ('y', np.float32,
@@ -197,6 +215,10 @@ class PeakPositions(strax.Plugin):
                  help='Maximum value for proximity values such as '
                       't_to_next_peak [ns]'))
 class PeakProximity(strax.OverlapWindowPlugin):
+    """
+    Look for peaks around a peak to determine how many peaks are in
+    proximity (in time) of a peak.
+    """
     depends_on = ('peak_basics',)
     dtype = [
         ('n_competing', np.int32,